Change LocationForecastResponse Geometry coordinates List<Float> to List<Int>

+++ project/app/src/main/java/no/uio/ifi/in2000/project/model/forecast/LocationForecastResponse.kt
@@ -13,7 +13,7 @@ data class LocationForecastResponse(
 @Serializable
 data class Geometry(
     val type: String,
-    val coordinates: List<Float>
+    val coordinates: List<Int>
 )
 //Comments Comments Comments
 @Serializable
